Provide confirmation fixes

+++ confirmation/main/main.go
@@ -5,8 +5,10 @@ import (
 	"confirmation/confirmation"
 	pb "confirmation/generated"
 	"confirmation/server"
+	"confirmation/store"
 	"context"
-	"github.com/go-redis/redis/v9"
+	"flag"
+	"fmt"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -16,27 +18,32 @@ import (
 )
 
 func main() {
-	cfg, _ := config.LoadConfig(".")
-	if err := os.RemoveAll("./generated"); err != nil {
+	productionMode := flag.Bool("production", false, "enable production mode")
+	flag.Parse()
+	cfgName := func() string {
+		if *productionMode {
+			return "prod"
+		}
+		return "dev"
+	}()
+	cfg, err := config.LoadConfig(cfgName)
+	if err != nil {
 		log.Fatal(err)
 	}
-	if err := os.Mkdir("./generated", os.ModePerm); err != nil {
-		log.Fatal(err)
-	}
-	if err := exec.Command("protoc", "--go_out=generated", "--go-grpc_out=generated", "--proto_path=proto", "proto/*.proto").Run(); err != nil {
-		log.Fatal(err)
-	}
-	redisClient := redis.NewClient(&redis.Options{Addr: cfg.RedisAddress})
-	if _, err := redisClient.Ping(context.Background()).Result(); err != nil {
-		log.Fatal(err)
-	}
-	defer func(redisClient *redis.Client) {
-		err := redisClient.Close()
-		if err != nil {
+	if !*productionMode {
+		if err := os.RemoveAll("./generated"); err != nil {
 			log.Fatal(err)
 		}
-	}(redisClient)
-	accountRepository := confirmation.NewRepository(redisClient)
+		if err := os.Mkdir("./generated", os.ModePerm); err != nil {
+			log.Fatal(err)
+		}
+		if err := exec.Command("protoc", "--go_out=generated", "--go-grpc_out=generated", "--proto_path=proto", "proto/*.proto").Run(); err != nil {
+			log.Fatal(err)
+		}
+	}
+	ctx := context.Background()
+	client := store.NewClient(ctx, fmt.Sprintf("%s:%s", cfg.RedisHostname, cfg.RedisPort))
+	accountRepository := confirmation.NewRepository(client)
 	accountUseCase := confirmation.NewUseCase(accountRepository)
 	accountService := confirmation.NewService(accountUseCase)
 	authInterceptor := server.NewInterceptor("Authorization", func(header string) error {
